+++ jqm-all/jqm-engine/src/main/java/com/enioka/jqm/engine/Helpers.java
@@ -114,7 +114,7 @@ final class Helpers
      * @param cnx
      *            the DbConn to use.
      */
-    static int createDeliverable(String path, String originalFileName, String fileFamily, Integer jobId, DbConn cnx)
+    static long createDeliverable(String path, String originalFileName, String fileFamily, Long jobId, DbConn cnx)
     {
         QueryResult qr = cnx.runUpdate("deliverable_insert", fileFamily, path, jobId, originalFileName, UUID.randomUUID().toString());
         return qr.getGeneratedId();
