+++ jqm-all/jqm-integration-tests/src/test/java/com/enioka/jqm/integration/tests/DeliverableTest.java
@@ -45,7 +45,7 @@ public class DeliverableTest extends JqmBaseTest
         GlobalParameter.setParameter(cnx, "enableWsApiSsl", "false");
         cnx.commit();
 
-        int id = JqmSimpleTest.create(cnx, "pyl.EngineApiSendDeliverable").addDefParameter("filepath", TestHelpers.node.getDlRepo())
+        Long id = JqmSimpleTest.create(cnx, "pyl.EngineApiSendDeliverable").addDefParameter("filepath", TestHelpers.node.getDlRepo())
                 .addDefParameter("fileName", "jqm-test-deliverable1.txt").run(this);
 
         List<InputStream> tmp = jqmClient.getJobDeliverablesContent(id);
@@ -67,7 +67,7 @@ public class DeliverableTest extends JqmBaseTest
         GlobalParameter.setParameter(cnx, "enableWsApiSsl", "false");
         cnx.commit();
 
-        int jobId = JqmSimpleTest.create(cnx, "pyl.EngineApiSendDeliverable").addDefParameter("filepath", TestHelpers.node.getDlRepo())
+        Long jobId = JqmSimpleTest.create(cnx, "pyl.EngineApiSendDeliverable").addDefParameter("filepath", TestHelpers.node.getDlRepo())
                 .addDefParameter("fileName", "jqm-test-deliverable2.txt").run(this);
 
         File f = new File(TestHelpers.node.getDlRepo() + "jqm-test-deliverable2.txt");
@@ -96,7 +96,7 @@ public class DeliverableTest extends JqmBaseTest
         GlobalParameter.setParameter(cnx, "enableWsApiSsl", "false");
         cnx.commit();
 
-        int jobId = JqmSimpleTest.create(cnx, "pyl.EngineApiSendDeliverable").addDefParameter("filepath", TestHelpers.node.getDlRepo())
+        Long jobId = JqmSimpleTest.create(cnx, "pyl.EngineApiSendDeliverable").addDefParameter("filepath", TestHelpers.node.getDlRepo())
                 .addDefParameter("fileName", "jqm-test-deliverable3.txt").run(this);
 
         File f = new File(TestHelpers.node.getDlRepo() + "jqm-test-deliverable3.txt");
@@ -129,7 +129,7 @@ public class DeliverableTest extends JqmBaseTest
         JqmDbClientFactory.setProperty("com.enioka.jqm.ws.truststorePass", "SuperPassword");
         jqmClient = JqmDbClientFactory.getClient();
 
-        int jobId = JqmSimpleTest.create(cnx, "pyl.EngineApiSendDeliverable").addDefParameter("filepath", TestHelpers.node.getDlRepo())
+        Long jobId = JqmSimpleTest.create(cnx, "pyl.EngineApiSendDeliverable").addDefParameter("filepath", TestHelpers.node.getDlRepo())
                 .addDefParameter("fileName", "jqm-test-deliverable4.txt").run(this);
 
         File f = new File(TestHelpers.node.getDlRepo() + "jqm-test-deliverable4.txt");
@@ -152,7 +152,7 @@ public class DeliverableTest extends JqmBaseTest
     @Test
     public void testGetAllDeliverables() throws Exception
     {
-        int jobId = JqmSimpleTest.create(cnx, "pyl.EngineApiSendDeliverable").addDefParameter("filepath", TestHelpers.node.getDlRepo())
+        Long jobId = JqmSimpleTest.create(cnx, "pyl.EngineApiSendDeliverable").addDefParameter("filepath", TestHelpers.node.getDlRepo())
                 .addDefParameter("fileName", "jqm-test-deliverable5.txt").run(this);
 
         List<com.enioka.jqm.client.api.Deliverable> tmp = jqmClient.getJobDeliverables(jobId);
