+++ jqm-all/jqm-integration-tests/src/test/java/com/enioka/jqm/integration/tests/MiscTest.java
@@ -190,7 +190,7 @@ public class MiscTest extends JqmBaseTest
                 null, "Franquin", "ModuleMachin", "other", "other", false, cnx);
 
         /// Create a running job that should be cleaned at startup
-        int i1 = jqmClient.enqueue("jqm-test-em", "test");
+        Long i1 = jqmClient.enqueue("jqm-test-em", "test");
         cnx.runUpdate("ji_update_status_by_id", TestHelpers.node.getId(), i1);
         cnx.runUpdate("jj_update_run_by_id", i1);
         cnx.commit();
@@ -209,7 +209,7 @@ public class MiscTest extends JqmBaseTest
                 null, "Franquin", "ModuleMachin", "other", "other", false, cnx);
 
         // Create a running job that should be cleaned at startup
-        int i = jqmClient.enqueue("jqm-test-em", "test");
+        Long i = jqmClient.enqueue("jqm-test-em", "test");
         cnx.runUpdate("ji_update_status_by_id", TestHelpers.node.getId(), i);
         cnx.commit();
