+++ jqm-all/jqm-integration-tests/src/test/java/com/enioka/jqm/integration/tests/NoApiPayloadTest.java
@@ -97,7 +97,7 @@ public class NoApiPayloadTest extends JqmBaseTest
     @Test
     public void testDisabledPayload() throws Exception
     {
-        int jd = CreationTools.createJobDef(null, true, "App", null, "jqm-tests/jqm-test-datetimemaven/target/test.jar", TestHelpers.qVip,
+        Long jd = CreationTools.createJobDef(null, true, "App", null, "jqm-tests/jqm-test-datetimemaven/target/test.jar", TestHelpers.qVip,
                 42, "MarsuApplication", null, "Franquin", "ModuleMachin", "other", "other", true, cnx);
         cnx.runUpdate("jd_update_set_enabled_by_id", false, jd);
         cnx.commit();
@@ -116,7 +116,7 @@ public class NoApiPayloadTest extends JqmBaseTest
     public void testMainTypeInjectWithFullApi() throws Exception
     {
         // Here, engine API + full API mix.
-        int i = JqmSimpleTest.create(cnx, "pyl.EngineApiInject", "jqm-test-pyl-hibapi").setSessionId("123X").expectOk(3).run(this);
+        Long i = JqmSimpleTest.create(cnx, "pyl.EngineApiInject", "jqm-test-pyl-hibapi").setSessionId("123X").expectOk(3).run(this);
 
         Assert.assertEquals(1, jqmClient.getJob(i).getMessages().size()); // 1 message per run created by payload
         Assert.assertEquals(100, (int) jqmClient.getJob(i).getProgress());
