+++ jqm-all/jqm-runner/jqm-runner-shell/src/main/java/com/enioka/jqm/runner/shell/ShellJobInstanceTracker.java
@@ -271,7 +271,7 @@ class ShellJobInstanceTracker implements JobInstanceTracker, ShellJobInstanceTra
     }
 
     @Override
-    public Integer getId()
+    public Long getId()
     {
         return this.ji.getId();
     }
