+++ jqm-all/jqm-ws/src/main/java/com/enioka/jqm/ws/api/ServiceClient.java
@@ -30,6 +30,7 @@ import com.enioka.jqm.client.api.Query;
 import com.enioka.jqm.client.api.Queue;
 import com.enioka.jqm.client.api.QueueStatus;
 import com.enioka.jqm.client.api.State;
+import com.enioka.jqm.client.jdbc.api.JqmClientFactory;
 import com.enioka.jqm.client.shared.JobRequestBaseImpl;
 import com.enioka.jqm.client.shared.QueryBaseImpl;
 import com.enioka.jqm.client.shared.SelfDestructFileStream;
@@ -88,7 +89,7 @@ public class ServiceClient
             target.startHeld();
         }
 
-        int i = target.enqueue();
+        long i = target.enqueue();
 
         JobInstance ji = new JobInstance();
         ji.setId(i);
@@ -115,68 +116,68 @@ public class ServiceClient
 
     @Path("ji/{id}")
     @POST
-    public int enqueueFromHistory(@PathParam("id") int jobIdToCopy)
+    public Long enqueueFromHistory(@PathParam("id") long jobIdToCopy)
     {
         return Helpers.getClient().enqueueFromHistory(jobIdToCopy);
     }
 
     @Path("ji/cancelled/{jobId}")
     @POST
-    public void cancelJob(@PathParam("jobId") int jobId)
+    public void cancelJob(@PathParam("jobId") long jobId)
     {
         Helpers.getClient().cancelJob(jobId);
     }
 
     @Path("ji/waiting/{jobId}")
     @DELETE
-    public void deleteJob(@PathParam("jobId") int jobId)
+    public void deleteJob(@PathParam("jobId") long jobId)
     {
         Helpers.getClient().deleteJob(jobId);
     }
 
     @Path("ji/killed/{jobId}")
     @POST
-    public void killJob(@PathParam("jobId") int jobId)
+    public void killJob(@PathParam("jobId") long jobId)
     {
         Helpers.getClient().killJob(jobId);
     }
 
     @Path("schedule/{scheduleId}")
     @DELETE
-    public void removeRecurrence(@PathParam("scheduleId") int scheduleId)
+    public void removeRecurrence(@PathParam("scheduleId") long scheduleId)
     {
         Helpers.getClient().removeRecurrence(scheduleId);
     }
 
     @Path("ji/paused/{jobId}")
     @POST
-    public void pauseQueuedJob(@PathParam("jobId") int jobId)
+    public void pauseQueuedJob(@PathParam("jobId") long jobId)
     {
         Helpers.getClient().pauseQueuedJob(jobId);
     }
 
     @Path("ji/paused/{jobId}")
     @DELETE
-    public void resumeQueuedJob(@PathParam("jobId") int jobId)
+    public void resumeQueuedJob(@PathParam("jobId") long jobId)
     {
         Helpers.getClient().resumeQueuedJob(jobId);
     }
 
-    public void resumeJob(@PathParam("jobId") int jobId)
+    public void resumeJob(@PathParam("jobId") long jobId)
     {
         resumeQueuedJob(jobId);
     }
 
     @Path("ji/running/paused/{jobId}")
     @POST
-    public void pauseRunningJob(@PathParam("jobId") int jobId)
+    public void pauseRunningJob(@PathParam("jobId") long jobId)
     {
         Helpers.getClient().pauseRunningJob(jobId);
     }
 
     @Path("ji/running/paused/{jobId}")
     @DELETE
-    public void resumeRunningJob(@PathParam("jobId") int jobId)
+    public void resumeRunningJob(@PathParam("jobId") long jobId)
     {
         Helpers.getClient().resumeRunningJob(jobId);
     }
@@ -190,48 +191,48 @@ public class ServiceClient
     @Path("ji/crashed/{jobId}")
     @Produces({ MediaType.APPLICATION_XML, MediaType.APPLICATION_JSON })
     @DELETE
-    public JobInstance restartCrashedJobObject(@PathParam("jobId") int jobId)
+    public JobInstance restartCrashedJobObject(@PathParam("jobId") long jobId)
     {
-        int i = Helpers.getClient().restartCrashedJob(jobId);
+        long i = Helpers.getClient().restartCrashedJob(jobId);
         return getJob(i);
     }
 
     @Path("q/{queueId: [0-9]+}/{jobId: [0-9]+}")
     @POST
-    public void setJobQueue(@PathParam("jobId") int jobId, @PathParam("queueId") int queueId)
+    public void setJobQueue(@PathParam("jobId") long jobId, @PathParam("queueId") long queueId)
     {
         Helpers.getClient().setJobQueue(jobId, queueId);
     }
 
     // No need to expose. Client side work.
 
-    public void setJobQueue(int jobId, Queue queue)
+    public void setJobQueue(Long jobId, Queue queue)
     {
         Helpers.getClient().setJobQueue(jobId, queue);
     }
 
     @POST
     @Path("ji/{jobId}/position/{newPosition}")
-    public void setJobQueuePosition(@PathParam("jobId") int jobId, @PathParam("newPosition") int newPosition)
+    public void setJobQueuePosition(@PathParam("jobId") long jobId, @PathParam("newPosition") int newPosition)
     {
         Helpers.getClient().setJobQueuePosition(jobId, newPosition);
     }
 
     @POST
     @Path("ji/{jobId}/priority/{priority}")
-    public void setJobPriority(@PathParam("jobId") int jobId, @PathParam("priority") int priority)
+    public void setJobPriority(@PathParam("jobId") long jobId, @PathParam("priority") int priority)
     {
         Helpers.getClient().setJobPriority(jobId, priority);
     }
 
-    public void setJobRunAfter(@PathParam("jobId") int jobId, @PathParam("whenToRun") Calendar whenToRun)
+    public void setJobRunAfter(@PathParam("jobId") long jobId, @PathParam("whenToRun") Calendar whenToRun)
     {
         Helpers.getClient().setJobRunAfter(jobId, whenToRun);
     }
 
     @POST
     @Path("ji/{jobId}/delay/{whenToRun}")
-    public void setJobRunAfter(@PathParam("jobId") int jobId, @PathParam("whenToRun") long whenToRun)
+    public void setJobRunAfter(@PathParam("jobId") long jobId, @PathParam("whenToRun") long whenToRun)
     {
         Calendar c = Calendar.getInstance();
         c.setTimeInMillis(whenToRun);
@@ -240,21 +241,21 @@ public class ServiceClient
 
     @POST
     @Path("schedule/{scheduleId}/queue/{queueId}")
-    public void setScheduleQueue(@PathParam("scheduleId") int scheduleId, @PathParam("queueId") int queueId)
+    public void setScheduleQueue(@PathParam("scheduleId") long scheduleId, @PathParam("queueId") long queueId)
     {
         Helpers.getClient().setScheduleQueue(scheduleId, queueId);
     }
 
     @POST
     @Path("schedule/{scheduleId}/cron/{cronExpression}")
-    public void setScheduleRecurrence(@PathParam("scheduleId") int scheduleId, @PathParam("cronExpression") String cronExpression)
+    public void setScheduleRecurrence(@PathParam("scheduleId") long scheduleId, @PathParam("cronExpression") String cronExpression)
     {
         Helpers.getClient().setScheduleRecurrence(scheduleId, cronExpression);
     }
 
     @POST
     @Path("schedule/{scheduleId}/priority/{priority}")
-    public void setSchedulePriority(@PathParam("scheduleId") int scheduleId, @PathParam("priority") int priority)
+    public void setSchedulePriority(@PathParam("scheduleId") long scheduleId, @PathParam("priority") int priority)
     {
         Helpers.getClient().setSchedulePriority(scheduleId, priority);
     }
@@ -263,7 +264,7 @@ public class ServiceClient
     @Produces({ MediaType.APPLICATION_XML, MediaType.APPLICATION_JSON })
     @Path("ji/{jobId}")
     @HttpCache("public, max-age=60")
-    public JobInstance getJob(@PathParam("jobId") int jobId)
+    public JobInstance getJob(@PathParam("jobId") long jobId)
     {
         return Helpers.getClient().getJob(jobId);
     }
@@ -342,14 +343,14 @@ public class ServiceClient
     @GET
     @Produces({ MediaType.APPLICATION_XML, MediaType.APPLICATION_JSON })
     @HttpCache("public, max-age=60")
-    public List<String> getJobMessages(@PathParam("jobId") int jobId)
+    public List<String> getJobMessages(@PathParam("jobId") long jobId)
     {
         return Helpers.getClient().getJobMessages(jobId);
     }
 
     // Not exposed. Use getJob => progress
 
-    public int getJobProgress(int jobId)
+    public int getJobProgress(long jobId)
     {
         throw new NotSupportedException();
     }
@@ -358,7 +359,7 @@ public class ServiceClient
     @GET
     @Produces({ MediaType.APPLICATION_XML, MediaType.APPLICATION_JSON })
     @HttpCache("public, max-age=60")
-    public List<Deliverable> getJobDeliverables(@PathParam("jobId") int jobId)
+    public List<Deliverable> getJobDeliverables(@PathParam("jobId") long jobId)
     {
         return Helpers.getClient().getJobDeliverables(jobId);
     }
@@ -366,7 +367,7 @@ public class ServiceClient
     // Not exposed. Returning a list of files is a joke anyway... Loop should be
     // client-side.
 
-    public List<InputStream> getJobDeliverablesContent(int jobId)
+    public List<InputStream> getJobDeliverablesContent(long jobId)
     {
         throw new NotSupportedException();
     }
@@ -385,7 +386,7 @@ public class ServiceClient
     @Path("ji/files/{id}")
     @Produces("application/octet-stream")
     @GET
-    public InputStream getDeliverableContent(@PathParam("id") int delId, @Context HttpServletResponse res)
+    public InputStream getDeliverableContent(@PathParam("id") long delId, @Context HttpServletResponse res)
     {
         SelfDestructFileStream fs = (SelfDestructFileStream) Helpers.getClient().getDeliverableContent(delId);
         res.setHeader("Content-Disposition", "attachment; filename=" + fs.nameHint);
@@ -395,7 +396,7 @@ public class ServiceClient
     @Path("ji/{jobId}/stderr")
     @Produces("application/octet-stream")
     @GET
-    public InputStream getJobLogStdErr(@PathParam("jobId") int jobId, @Context HttpServletResponse res)
+    public InputStream getJobLogStdErr(@PathParam("jobId") long jobId, @Context HttpServletResponse res)
     {
         SelfDestructFileStream fs = (SelfDestructFileStream) Helpers.getClient().getJobLogStdErr(jobId);
         res.setHeader("Content-Disposition", "attachment; filename=" + fs.nameHint);
@@ -405,7 +406,7 @@ public class ServiceClient
     @Path("ji/{jobId}/stdout")
     @Produces("application/octet-stream")
     @GET
-    public InputStream getJobLogStdOut(@PathParam("jobId") int jobId, @Context HttpServletResponse res)
+    public InputStream getJobLogStdOut(@PathParam("jobId") long jobId, @Context HttpServletResponse res)
     {
         SelfDestructFileStream fs = (SelfDestructFileStream) Helpers.getClient().getJobLogStdOut(jobId);
         res.setHeader("Content-Disposition", "attachment; filename=" + fs.nameHint);
@@ -428,7 +429,7 @@ public class ServiceClient
 
     @Path("q/{qId}/pause")
     @POST
-    public void pauseQueue(@PathParam("qId") int qId)
+    public void pauseQueue(@PathParam("qId") long qId)
     {
         Queue q = new Queue();
         q.setId(qId);
@@ -442,7 +443,7 @@ public class ServiceClient
 
     @Path("q/{qId}/pause")
     @DELETE
-    public void resumeQueue(@PathParam("qId") int qId)
+    public void resumeQueue(@PathParam("qId") long qId)
     {
         Queue q = new Queue();
         q.setId(qId);
@@ -456,7 +457,7 @@ public class ServiceClient
 
     @Path("q/{qId}/clear")
     @POST
-    public void clearQueue(@PathParam("qId") int qId)
+    public void clearQueue(@PathParam("qId") long qId)
     {
         Queue q = new Queue();
         q.setId(qId);
@@ -465,7 +466,7 @@ public class ServiceClient
 
     @Path("q/{qId}/status")
     @GET
-    public QueueStatus getQueueStatus(@PathParam("qId") int qId)
+    public QueueStatus getQueueStatus(@PathParam("qId") long qId)
     {
         Queue q = new Queue();
         q.setId(qId);
@@ -479,7 +480,7 @@ public class ServiceClient
 
     @Path("q/{qId}/enabled-capacity")
     @GET
-    public int getQueueEnabledCapacity(@PathParam("qId") int qId)
+    public int getQueueEnabledCapacity(@PathParam("qId") long qId)
     {
         Queue q = new Queue();
         q.setId(qId);
